+++ Benchmarks/Program.cs
@@ -1,4 +1,6 @@
 ﻿using BenchmarkDotNet.Running;
 using Benchmarks;
 
+// Run all benchmarks
 BenchmarkRunner.Run<GetModelType>();
+BenchmarkRunner.Run<GetMerakiApiDomain>();